perf(create-project): move init store to base store
