+++ packages/create-project/src/bin/index.js
@@ -4,7 +4,6 @@
 import path from 'path';
 
 import chalk from 'chalk';
-import debug from 'debug';
 
 import { handleUnhandledRejection } from '@cat-org/utils';
 
@@ -12,7 +11,6 @@ import logger from 'utils/logger';
 import cliOptions from 'utils/cliOptions';
 import validateProject from 'utils/validateProject';
 import base from 'stores/base';
-import type StoreType from 'stores';
 
 handleUnhandledRejection();
 
@@ -29,19 +27,7 @@ handleUnhandledRejection();
     )}}`,
   );
 
-  const storeNames = [];
-  const stores = (await base.run(ctx)).filter(
-    ({ constructor: { name } }: StoreType): boolean => {
-      if (storeNames.includes(name)) return false;
-
-      storeNames.push(name);
-      return true;
-    },
-  );
-
-  debug('create-project:bin')(stores);
-  for (const store of stores) await store.end(ctx);
-  await base.end(ctx);
+  await base.init(ctx);
 
   logger.succeed('Done');
 })();
