+++ packages/create-project/src/stores/base.js
@@ -1,5 +1,7 @@
 // @flow
 
+import debug from 'debug';
+
 import pkg from './pkg';
 import gitignore from './gitignore';
 import npmignore from './npmignore';
@@ -9,17 +11,35 @@ import license from './license';
 import readme from './readme';
 import Store from './index';
 
+const debugLog = debug('create-project:store:base');
+
 /** base store */
 class Base extends Store {
   subStores = [pkg, gitignore, npmignore, configs, flow, license, readme];
 
   /**
    * @example
-   * bese.end(ctx)
+   * base.init(ctx)
    *
    * @param {Object} ctx - store context
    */
-  end = async ({ pkg: { repository } = {} }: $PropertyType<Store, 'ctx'>) => {
+  init = async (ctx: $PropertyType<Store, 'ctx'>) => {
+    const storeNames = [];
+    const stores = (await this.run(ctx)).filter(
+      ({ constructor: { name } }: Store): boolean => {
+        if (storeNames.includes(name)) return false;
+
+        storeNames.push(name);
+        return true;
+      },
+    );
+
+    debugLog(stores);
+
+    for (const store of stores) await store.end(ctx);
+
+    const { pkg: { repository } = {} } = ctx;
+
     await this.execa(
       'yarn flow-typed install',
       'git add .',
